valuable_code_copy/pyproject.toml
[project]
name = "financial-advisor-assistant"
version = "0.1.0"
description = "Financial Advisor Life Insurance Assistant - Multi-agent RAG system with confidence-based routing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=0.3.14",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "ragas>=0.2.10",
    "chainlit>=1.0.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
    "qdrant-client>=1.7.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
    "cohere>=5.12.0",
    "rapidfuzz>=3.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.0",
    "langchain-experimental>=0.0.50",
    "pillow>=11.3.0",
    "unstructured>=0.18.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "retrieval", "evaluation", "utils", "config"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
] 
